Make it clear what QuerySummary is
Bug: n/a
Test: n/a

AOSP: 8053b5eab71ad04b32fc5102b1aa5f99333f3ce7
