+++ base/src/com/google/idea/blaze/base/qsync/BazelQueryRunner.java
@@ -28,6 +28,7 @@ import com.google.idea.blaze.common.PrintOutput;
 import com.google.idea.blaze.exception.BuildException;
 import com.google.idea.blaze.qsync.query.QuerySpec;
 import com.google.idea.blaze.qsync.query.QuerySummary;
+import com.google.idea.blaze.qsync.query.QuerySummaryImpl;
 import com.google.idea.common.experiments.BoolExperiment;
 import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.project.Project;
@@ -133,7 +134,7 @@ public class BazelQueryRunner implements QueryRunner {
     logger.info(String.format("Summarising query from %s", in));
     Instant start = Instant.now();
     try {
-      QuerySummary summary = QuerySummary.create(queryStrategy, in);
+      QuerySummary summary = QuerySummaryImpl.create(queryStrategy, in);
       logger.info(
           String.format(
               "Summarised query in %ds", Duration.between(start, Instant.now()).toSeconds()));
